Split staging duties Fixes #20
ok, so a bit of a major whack.

Deep down in mission.common.sequencer.scheduler we now have a
flag indicating that we want to stage. This flag can be set by mission
code in a SYNC function (so it can be scheduled), and the scheduler
knows to look at it and trigger the ASYNC stage call when needed.

with this in place, we can now allow the Scheduler to handle
looping, triggering the Sequencer to run the mission phases.