add STAGER
associated tweaks:
- tinker activates stager
- tweaks to tinker mission plan
- common proxies throttle methods to sequencer
- sequencer proxies throttle methods to scheduler
- scheduler owns the throttle
this allows me to tinker with ways to have the scheduler DIP
the throttle while staging, since we have to actually perform
the staging operation within the scheduler loop.

I may pull the staging-and-throttle cruft out into a separate
module that the scheduler will use, but that linkage is to
an ASYNC operation so no actual indirection is possible.